requirements.txt
pandas==2.2.2
pyarrow==17.0.0
scikit-learn==1.5.1
lightgbm==4.5.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
boto3==1.35.2
pytest==8.3.3
requests==2.25.1